Some refactoring and fixes to make it easier to use the same code for full extraction and updates to medical terminologies
